avance wiew con error delete

+++ app/controllers/services_controller.rb
@@ -1,11 +1,17 @@
 class ServicesController < ApplicationController
-  before_action :service_params, only: %i[create update]
+  #skip_before_action :authenticate_user!, only: :profile
+  before_action :service_params, only: %i[create update destroy]
   before_action :set_service, only: %i[show edit update destroy]
-  before_action :set_user, only: %i[new create update]
+  before_action :set_user, only: %i[new create update destroy]
 
   def index
     @services = policy_scope(Service)
   end
+
+  def profile
+    @my_services = current_user.services
+    authorize @my_services
+  end  
 
   def new
     @service = Service.new
@@ -38,12 +44,15 @@ class ServicesController < ApplicationController
   end
 
   def destroy
+    authorize @service
     if @service.destroy
-      redirect_to services_path
+      redirect_to service_path
     else
       render :show
     end
   end
+
+ 
 
   private
